+++ packages/annotation/src/misc/index.tsx
@@ -1,3 +1,4 @@
 export * from './Connector'
 export * from './GridStripes'
+export * from './Stripe'
 export * from './types'