+++ packages/annotation/src/misc/types.ts
@@ -1,9 +1,11 @@
-import { GridLinesProps, LineProps } from '@chsk/core'
-
-export interface GridStripesProps extends GridLinesProps {
-    /** toggle intervals used for stripes */
-    parity?: 'even' | 'odd'
-}
+import {
+    GridLinesProps,
+    LineProps,
+    NumericPositionSpec,
+    PositionSpec,
+    PositionUnits,
+    SvgElementVariantProps,
+} from '@chsk/core'
 
 export interface ConnectorProps extends LineProps {
     /** variant */
@@ -14,4 +16,20 @@ export interface ConnectorProps extends LineProps {
     elbow?: number
     /** unit for elbow position */
     elbowUnit?: 'absolute' | 'relative'
+}
+
+export interface GridStripesProps extends GridLinesProps {
+    /** toggle intervals used for stripes */
+    parity?: 'even' | 'odd'
+}
+
+export interface StripeProps extends SvgElementVariantProps, Pick<GridLinesProps, 'expansion'> {
+    /** variant */
+    variant: 'x' | 'y'
+    /** domain interval */
+    domain: PositionSpec
+    /** units for domain */
+    domainUnits?: PositionUnits
+    /** shifts for domain boundaries (multiples of bandwidth) */
+    shift?: NumericPositionSpec
 }